4540 - added types of variables, removed comments from phtml files, provided small changes

+++ app/code/local/Tim/Recommendation/Helper/Data.php
@@ -23,12 +23,12 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
     }
 
     /**
-     * @param string|int $customerId
+     * @param int $customerId
      * @return string
      */
     public function getCustomerUserTypeId($customerId)
     {
-        $user = Mage::getModel('tim_recommendation/user')->load($customerId, 'customer_id');
+        $user = Mage::getModel('tim_recommendation/user')->load((int) $customerId, 'customer_id');
         $userTypeId = $user->getUserType();
 
         return $userTypeId;
@@ -47,7 +47,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
      */
     public function getRecommendationMd5($recomId)
     {
-        $object = Mage::getModel('tim_recommendation/recommendation')->load($recomId);
+        $object = Mage::getModel('tim_recommendation/recommendation')->load((int) $recomId);
         $md5hash = $object->getMd5();
 
         return $md5hash;
@@ -58,7 +58,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
      */
     public function getSiteUrl()
     {
-        $customerId = Mage::helper('customer')->getCustomer()->getEntityId();
+        $customerId = (int) Mage::helper('customer')->getCustomer()->getEntityId();
         $siteUrl = Mage::getModel('tim_recommendation/user')->load($customerId, 'customer_id')->getWww();
         return $siteUrl;
     }
@@ -72,9 +72,9 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
     {
         if (empty($customerId)) {
             $customerId = Mage::helper('customer')->getCustomer()->getEntityId();
-            $nick = Mage::getModel('tim_recommendation/user')->load($customerId, 'customer_id')->getNick();
+            $nick = Mage::getModel('tim_recommendation/user')->load((int) $customerId, 'customer_id')->getNick();
         } else {
-            $nick = Mage::getModel('tim_recommendation/user')->load($customerId, 'customer_id')->getNick();
+            $nick = Mage::getModel('tim_recommendation/user')->load((int) $customerId, 'customer_id')->getNick();
         }
         return $nick;
     }
@@ -84,7 +84,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
      */
     public function getCustomerBanner()
     {
-        $customerId = Mage::helper('customer')->getCustomer()->getEntityId();
+        $customerId = (int) Mage::helper('customer')->getCustomer()->getEntityId();
         $banner = Mage::getModel('tim_recommendation/user')->load($customerId, 'customer_id')->getAd();
         if (!empty($banner)) {
             $banner = Mage::getBaseUrl(Mage_Core_Model_Store::URL_TYPE_WEB) . $banner;
@@ -99,7 +99,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
      */
     public function getCustomerAvatar()
     {
-        $customerId = Mage::helper('customer')->getCustomer()->getEntityId();
+        $customerId = (int) Mage::helper('customer')->getCustomer()->getEntityId();
         $avatar = Mage::getModel('tim_recommendation/user')->load($customerId, 'customer_id')->getAvatar();
         if (!empty($avatar)) {
             $avatar = Mage::getBaseUrl(Mage_Core_Model_Store::URL_TYPE_WEB) . $avatar;
@@ -111,7 +111,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
 
     public function getCustomerDescription()
     {
-        $customerId = Mage::helper('customer')->getCustomer()->getEntityId();
+        $customerId = (int) Mage::helper('customer')->getCustomer()->getEntityId();
         $siteUrl = Mage::getModel('tim_recommendation/user')->load($customerId, 'customer_id')->getDescription();
         return $siteUrl;
     }
@@ -123,7 +123,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
      */
     public function getCustomerName($customerId)
     {
-        $customer = Mage::getModel('customer/customer')->load($customerId);
+        $customer = Mage::getModel('customer/customer')->load((int) $customerId);
         $name = $customer->getFirstname() . ' ' . $customer->getLastname();
 
         return $name;
@@ -136,7 +136,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
      */
     public function getCustomerNickname($customerId)
     {
-        $recommendationUser = Mage::getModel('tim_recommendation/user')->load($customerId, 'customer_id');
+        $recommendationUser = Mage::getModel('tim_recommendation/user')->load((int) $customerId, 'customer_id');
         $nickname = $recommendationUser->getNick();
 
         return $nickname;
@@ -149,9 +149,9 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
      */
     public function getCustomerNameOrNick($customerId)
     {
-        $customerName = $this->getCustomerNickname($customerId);
+        $customerName = $this->getCustomerNickname((int) $customerId);
         if (empty($customerName)) {
-            $customerName = $this->getCustomerName($customerId);
+            $customerName = $this->getCustomerName((int) $customerId);
         }
 
         return $customerName;
@@ -164,7 +164,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
      */
     public function getUserTypeName($userTypeId)
     {
-        $userType = Mage::getModel('tim_recommendation/userType')->load($userTypeId, 'user_type_id');
+        $userType = Mage::getModel('tim_recommendation/userType')->load((int) $userTypeId, 'user_type_id');
         $userTypeName = $userType->getName();
 
         return $userTypeName;
@@ -178,7 +178,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
     public function getOpinionMediaPath($opinionId)
     {
         $collection = Mage::getModel('tim_recommendation/media')->getCollection();
-        $collection->addFieldToFilter('recom_id', $opinionId);
+        $collection->addFieldToFilter('recom_id', (int) $opinionId);
         $data = $collection->getData();
         $mediaPaths = array();
         foreach ($data as $item) {
@@ -203,7 +203,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
         $opinionCollection->addFieldToSelect('user_id');
         $opinionCollection->getSelect()->where('parent IS NULL');
         if ($customerId) {
-            $opinionCollection->addFieldToFilter('user_id', $customerId);
+            $opinionCollection->addFieldToFilter('user_id', (int) $customerId);
         }
         $qty = count($opinionCollection->getData());
 
@@ -302,7 +302,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
      */
     public function checkForOpinionComment($recomId)
     {
-        $recommendationRow = Mage::getModel('tim_recommendation/recommendation')->load($recomId);
+        $recommendationRow = Mage::getModel('tim_recommendation/recommendation')->load((int) $recomId);
         if ($parent = $recommendationRow->getParent()) {
             return $parent;
         } else {
@@ -389,7 +389,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
     public function getAddOpinionPoint()
     {
         $points = Mage::getStoreConfig('tim_settings/customer_points/add_opinion');
-        return (int)$points;
+        return (int) $points;
     }
 
     /**
@@ -399,7 +399,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
     public function getAddOpinionImagePoint()
     {
         $points = Mage::getStoreConfig('tim_settings/customer_points/image_to_opinion');
-        return (int)$points;
+        return (int) $points;
     }
 
     /**
@@ -409,7 +409,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
     public function getAddOpinionMoviePoint()
     {
         $points = Mage::getStoreConfig('tim_settings/customer_points/movie_to_opinion');
-        return (int)$points;
+        return (int) $points;
     }
 
     /**
@@ -419,7 +419,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
     public function getAddComentPoint()
     {
         $points = Mage::getStoreConfig('tim_settings/customer_points/add_comment');
-        return (int)$points;
+        return (int) $points;
     }
 
     /**
@@ -456,8 +456,8 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
      */
     public function getUserScore($customerId)
     {
-        $customer = Mage::getModel('customer/customer')->load($customerId);
-        $user = Mage::getModel('tim_recommendation/user')->load($customerId, 'customer_id');
+        $customer = Mage::getModel('customer/customer')->load((int) $customerId);
+        $user = Mage::getModel('tim_recommendation/user')->load((int) $customerId, 'customer_id');
         $customerPoints = $user->getPoints();
         $customerLevel = $user->getLevel();
         $userLevelsClient = $this->getUserLevelClient();
@@ -503,7 +503,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
      */
     public function getImages($recomId)
     {
-        $opinionMedia = $this->getOpinionMediaPath($recomId);
+        $opinionMedia = $this->getOpinionMediaPath((int) $recomId);
         $images = array();
         foreach ($opinionMedia as $key => $value) {
             if ($key === 'url/youtube') {
@@ -522,8 +522,8 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
     {
         //check on acceptance for opinion or comment
         if (!$recommendationModel->getAcceptance()) {
-            $recomId = $recommendationModel->getRecomId();
-            $userId = $recommendationModel->getUserId();
+            $recomId = (int) $recommendationModel->getRecomId();
+            $userId = (int) $recommendationModel->getUserId();
             $userModel = Mage::getModel('tim_recommendation/user')->load($userId, 'customer_id');
             $mediaModel = Mage::getModel('tim_recommendation/media')->load($recomId, 'recom_id');
             $opinionOrComment = $this->checkOpinionOrComment($recomId);
@@ -558,7 +558,7 @@ class Tim_Recommendation_Helper_Data extends Mage_Core_Helper_Abstract
      */
     function checkOpinionOrComment($recomId)
     {
-        $parent = Mage::getModel('tim_recommendation/recommendation')->load($recomId)->getParent();
+        $parent = Mage::getModel('tim_recommendation/recommendation')->load((int) $recomId)->getParent();
         if ($parent) {
             return 'comment';
         } else {
